fix: change sessionRecordingSampleRate from double to int
Sample rate is a percentage (0-100) and both native platforms expect Int.

+++ lib/cx_session_replay_options.dart
@@ -2,7 +2,7 @@
 class CXSessionReplayOptions {
   final double captureScale;
   final double captureCompressQuality;
-  final double sessionRecordingSampleRate;
+  final int sessionRecordingSampleRate;
   final bool autoStartSessionRecording;
   final bool? maskAllTexts;
   final List<String>? textsToMask;
